chore: webhook consumer updated with preload information
