+++ apps/api/bin/docker-entrypoint-webhook-consumer.sh
@@ -7,6 +7,6 @@ python manage.py wait_for_migrations
 
 
 # Run the processes
-python manage.py webhook_consumer \
+python manage.py run_webhook_consumer \
   --queue ${WEBHOOK_QUEUE_NAME:-plane.webhook} \
   --prefetch ${WEBHOOK_PREFETCH_COUNT:-10}
